requirements.txt
fastapi
requests
jsonlines
python-dotenv
uvicorn
pandas
